Added PageTabSearch
